refactor: extract safegaurds outside release plans
